tirando o arquvio do app para ter navagação #10
