+++ src/App.jsx
@@ -4,93 +4,13 @@ import TodoForm from "./components/TodoForm";
 import "./App.css";
 import Search from "./components/Search";
 import Filter from "./components/filter";
+import TodoList from "./components/ToDoList";
 
 function App() {
-  const [todos, setTodos] = useState([
-    {
-      id: 1,
-      text: "criar funcionalidade x no sistema",
-      category: "trabalho",
-      isComplete: false,
-    },
-    {
-      id: 2,
-      text: "Iracing",
-      category: "Pessoal",
-      isComplete: false,
-    },
-    {
-      id: 3,
-      text: "Subir para o git",
-      category: "trabalho",
-      isComplete: false,
-    },
-  ]);
-  const [search, setSearch] = useState("");
-
-  const addTodo = (text, category) => {
-    const newTodos = [
-      ...todos,
-      {
-        id: Math.floor(Math.random() * 1000),
-        text,
-        category,
-        isComplete: false,
-      },
-    ];
-    setTodos(newTodos);
-  };
-  const removeTodo = (id) => {
-    const newTodos = [...todos];
-    const filteredTodos = newTodos.filter((todo) =>
-      todo.id !== id ? Todo : null
-    );
-    setTodos(filteredTodos);
-  };
-  const competeTodo = (id) => {
-    const newTodos = [...todos];
-    newTodos.map((todo) =>
-      todo.id === id ? (todo.isComplete = !todo.isComplete) : todo
-    );
-    setTodos(newTodos);
-  };
-  const [filter, setFilter] = useState("All");
-  const [sort, setSort] = useState("Asc");
-
   return (
     <div className="app">
-      <h1>Lista de Tarefas</h1>
-      <Search search={search} setSearch={setSearch} />
-      <Filter filter={filter} setFilter={setFilter} setSort={setSort} />
-      <div className="todo-list">
-        {todos
-          .filter((todo) =>
-            filter === "All"
-              ? true
-              : filter === "Complete"
-              ? todo.isComplete
-              : !todo.isComplete
-          )
-          .filter((todo) =>
-            todo.text.toLowerCase().includes(search.toLowerCase())
-          )
-          .sort((a, b) =>
-            sort === "Asc"
-              ? a.text.localeCompare(b.text)
-              : b.text.localeCompare(a.text)
-          )
-          .map((todo) => (
-            <Todo
-              key={todo.id}
-              todo={todo}
-              removeTodo={removeTodo}
-              competeTodo={competeTodo}
-            />
-          ))}
-      </div>
-      <TodoForm addTodo={addTodo} />
+      <TodoList />
     </div>
   );
 }
-
 export default App;
